# Using Apiary client for isHnBucket checks 