--- a/gcsio/src/main/java/com/google/cloud/hadoop/gcsio/GoogleCloudStorageImpl.java
+++ b/gcsio/src/main/java/com/google/cloud/hadoop/gcsio/GoogleCloudStorageImpl.java
@@ -46,6 +46,7 @@ import com.google.api.gax.core.FixedCredentialsProvider;
 import com.google.api.services.storage.Storage;
 import com.google.api.services.storage.StorageRequest;
 import com.google.api.services.storage.model.Bucket;
+import com.google.api.services.storage.model.BucketStorageLayout;
 import com.google.api.services.storage.model.Buckets;
 import com.google.api.services.storage.model.ComposeRequest;
 import com.google.api.services.storage.model.Objects;
@@ -1824,7 +1825,7 @@ public class GoogleCloudStorageImpl implements GoogleCloudStorage {
     checkNotNull(listedFolder, "Must provide a non-null container for listedFolder.");
 
     ListFoldersPagedResponse listFolderRespose =
-        storageControlClient.listFolders(listFoldersRequest);
+        lazyGetStorageControlClient().listFolders(listFoldersRequest);
     try (ITraceOperation op = TraceOperation.addToExistingTrace("gcs.folders.list")) {
       Iterator<Folder> itemsIterator = listFolderRespose.getPage().getValues().iterator();
       while (itemsIterator.hasNext()) {
@@ -2414,20 +2415,11 @@ public class GoogleCloudStorageImpl implements GoogleCloudStorage {
       return isEnabled;
     }
 
-    String prefix = src.getPath().substring(1);
-
-    StorageControlClient storageControlClient = lazyGetStorageControlClient();
-    GetStorageLayoutRequest request =
-        GetStorageLayoutRequest.newBuilder()
-            .setPrefix(prefix)
-            .setName(StorageLayoutName.format("_", bucketName))
-            .build();
-
+    Storage.Buckets.GetStorageLayout request =
+        initializeRequest(storage.buckets().getStorageLayout(bucketName), bucketName);
     try (ITraceOperation to = TraceOperation.addToExistingTrace("getStorageLayout.HN")) {
-      StorageLayout storageLayout = storageControlClient.getStorageLayout(request);
-      boolean result =
-          storageLayout.hasHierarchicalNamespace()
-              && storageLayout.getHierarchicalNamespace().getEnabled();
+      BucketStorageLayout layout = request.execute();
+      boolean result = layout.getHierarchicalNamespace().getEnabled();
 
       logger.atInfo().log("Checking if %s is HN enabled returned %s", src, result);
 
@@ -2461,7 +2453,7 @@ public class GoogleCloudStorageImpl implements GoogleCloudStorage {
 
     try (ITraceOperation to = TraceOperation.addToExistingTrace("renameHnFolder")) {
       logger.atFine().log("Renaming HN folder (%s -> %s)", src, dst);
-      this.storageControlClient.renameFolderOperationCallable().call(request);
+      lazyGetStorageControlClient().renameFolderOperationCallable().call(request);
     } catch (Throwable t) {
       logger.atSevere().withCause(t).log("Renaming %s to %s failed", src, dst);
       throw t;